megano/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
diploma-frontend @ file:///Users/alexander/Documents/Programming/Learning/Skillbox/1.%20My%20professions/Profession%20Python%20developer/1.%20Basic%20courses/3.%20Python%20Django%20framework/Final%20work/python_django_diploma/diploma-frontend/dist/diploma-frontend-0.2.tar.gz
Django==4.1.7
djangorestframework==3.14.0
docutils==0.19
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
Pillow==9.4.0
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
